requirements.txt
fastapi==0.65.2
pymongo
pydantic[email]
python-decouple==3.3
uvicorn==0.11.8
